backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1
openai-whisper==20231117
ffmpeg-python==0.2.0
numpy==1.24.3
torch==2.1.0
torchaudio==2.1.0
deep-translator==1.11.4
beautifulsoup4==4.14.3
language-tool-python==2.8.1

# Speech-to-Text options (uncomment based on your choice):

# Option 1: OpenAI Whisper API
# openai==1.3.0

# Option 2: Local Whisper model
# openai-whisper==20231117
# torch==2.1.0

# Option 3: Google Cloud Speech-to-Text
# google-cloud-speech==2.21.0

# Option 4: AssemblyAI
# assemblyai==0.17.0

# Audio processing utilities
# pydub==0.25.1
# ffmpeg-python==0.2.0
